+++ CHANGELOG.md
@@ -64,5 +64,6 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 - Fix staging deploy workflow [#613](https://github.com/azavea/echo-locator/pull/613)
 - Fix staging basemap glyphs error [#707](https://github.com/azavea/echo-locator/pull/707)
+- Refactor text search desktop list view [#764](https://github.com/azavea/echo-locator/pull/764)
 
 ### Removed